# openapi/co.v1.yaml
openapi: 3.0.3
info:
  title: Scimigo Curriculum Orchestrator API
  version: 0.1.0
  description: |
    The Curriculum Orchestrator manages learning tracks, sessions, and orchestrates
    evaluations and tutoring for the SciMigo platform.
servers:
  - url: https://co.scimigo.com/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Local development
security:
  - bearerAuth: []
paths:
  /tracks:
    get:
      summary: List tracks
      tags: [Tracks]
      parameters:
        - in: query
          name: subject
          schema:
            type: string
            enum: [coding, math, systems]
          description: Filter by subject
        - in: query
          name: label
          schema:
            type: string
          description: Filter by label
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Track'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tracks/{id}:
    get:
      summary: Get a track by ID or slug
      tags: [Tracks]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Track ID (UUID) or slug
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        subject:
          type: string
          enum: [coding, math, systems]
        title:
          type: string
        modules:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
              outcomes:
                type: array
                items:
                  type: string
        labels:
          type: array
          items:
            type: string
        version:
          type: string
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...